+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Ledger.hs
@@ -88,10 +88,8 @@ import Cardano.Ledger.Shelley.LedgerState (
   LedgerState (..),
   UTxOState (..),
   asTreasuryL,
-  certVStateL,
   utxosGovStateL,
   utxosUtxoL,
-  vsCommitteeStateL,
  )
 import Cardano.Ledger.Shelley.Rules (
   LedgerEnv (..),
@@ -422,7 +420,7 @@ ledgerTransition = do
                   currentEpoch
                   pp
                   (govState ^. constitutionGovStateL . constitutionScriptL)
-                  (certState ^. certVStateL . vsCommitteeStateL)
+                  certState
               , proposals
               , govProcedures
               )