+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/Gov.hs
@@ -15,6 +15,7 @@ import Data.Foldable
 import Data.Coerce
 
 import Cardano.Ledger.BaseTypes
+import Cardano.Ledger.CertState
 import Cardano.Ledger.Conway.Governance
 import Cardano.Ledger.Conway.PParams
 import Cardano.Ledger.Conway.Rules
@@ -251,8 +252,12 @@ govProceduresSpec ge@GovEnv {..} ps =
         , geEpoch <= act ^. gasExpiresAfterL
         , f (gasAction act)
         ]
+      committeeState = vsCommitteeState (certVState geCertState)
+      knownDReps = Map.keysSet $ vsDReps (certVState geCertState)
+      knownStakePools = Map.keysSet $ psStakePoolParams (certPState geCertState)
+      knownCommitteeAuthorizations = authorizedHotCommitteeCredentials committeeState
       committeeVotableActionIds =
-        actions (isCommitteeVotingAllowed geEpoch geCommitteeState)
+        actions (isCommitteeVotingAllowed geEpoch committeeState)
       drepVotableActionIds =
         actions isDRepVotingAllowed
       stakepoolVotableActionIds =
@@ -263,14 +268,20 @@ govProceduresSpec ge@GovEnv {..} ps =
               forAll votingProcsMap $ \kvp ->
                 match kvp $ \voter mapActVote ->
                   (caseOn voter)
-                    ( branch $ \_c ->
-                        subset_ (dom_ mapActVote) (lit $ Set.fromList committeeVotableActionIds)
+                    ( branch $ \committeeHotCred ->
+                        [subset_ (dom_ mapActVote) (lit $ Set.fromList committeeVotableActionIds)
+                        , member_ committeeHotCred $ lit knownCommitteeAuthorizations
+                        ]
                     )
-                    ( branch $ \_c ->
-                        subset_ (dom_ mapActVote) (lit $ Set.fromList drepVotableActionIds)
+                    ( branch $ \drepCred ->
+                        [ subset_ (dom_ mapActVote) (lit $ Set.fromList drepVotableActionIds)
+                        , member_ drepCred $ lit knownDReps
+                        ]
                     )
-                    ( branch $ \_c ->
-                        subset_ (dom_ mapActVote) (lit $ Set.fromList stakepoolVotableActionIds)
+                    ( branch $ \poolKeyHash ->
+                        [ subset_ (dom_ mapActVote) (lit $ Set.fromList stakepoolVotableActionIds)
+                        , member_ poolKeyHash $ lit knownStakePools
+                        ]
                     )
           , forAll proposalProcs $ \proc ->
               match proc $ \deposit returnAddr govAction _ ->