+++ eras/conway/impl/CHANGELOG.md
@@ -8,6 +8,13 @@
 * Added `cgceCurrentCommittee` and `cgceCommitteeProposals` to `ConwayGovCertEnv`
 * Added `proposalsWithPurpose`, `isGovActionWithPurpose` and `ToGovActionPurpose`
 * Added `ConwayTxRefScriptsSizeTooBig` predicate failure to `ConwayLedgerPredFailure`
+* Replaced `geCommitteeState` with `geCertState` in `GovEnv`
+* Added `VotersDoNotExist` predicate failure to `ConwayGovPredFailure`
+
+### `testlib`
+
+* Change the return type of `resignCommitteeColdKey`
+* Add an argument to `registerCommitteeHotKeys`
 
 ## 1.15.1.0